adding feature to handle missing entry deletion error

+++ src/components/DeleteDataModal.tsx
@@ -20,6 +20,13 @@ export default function DeleteDataModal(props: any) {
   const [deleteSquat, setDeleteSquat] = useState<boolean>(false);
   const [deleteDeadlift, setDeleteDeadlift] = useState<boolean>(false);
 
+  const [missingBodyWeightEntry, setMissingBodyWeightEntry] =
+    useState<boolean>(false);
+  const [missingBenchEntry, setMissingBenchEntry] = useState<boolean>(false);
+  const [missingSquatEntry, setMissingSquatEntry] = useState<boolean>(false);
+  const [missingDeadliftEntry, setMissingDeadliftEntry] =
+    useState<boolean>(false);
+
   const supabase = useSupabaseClient();
 
   const errorToastOptions: ToastOptions = {
@@ -60,17 +67,38 @@ export default function DeleteDataModal(props: any) {
   ) {
     const { data, error } = await supabase
       .from("pr_data")
-      .delete()
+      .select("weight")
       .eq("user_id", user_id)
       .eq("date", date)
       .eq("exercise", exercise);
-    if (error) console.log(error);
+    if (data !== null && data.length > 1) {
+      const { data, error } = await supabase
+        .from("pr_data")
+        .delete()
+        .eq("user_id", user_id)
+        .eq("date", date)
+        .eq("exercise", exercise);
+      if (error) console.log(error);
+    } else {
+      toast.error(
+        "Looks like there's no record for a " + exercise + " PR on that day",
+        errorToastOptions
+      );
+      return true;
+    }
+    return false;
   }
 
   async function deleteData(e: { preventDefault: () => void }) {
     e.preventDefault();
     if (date === "") {
       toast.error("Please enter a date 📅", errorToastOptions);
+      return;
+    } else if (new Date(date) > new Date()) {
+      toast.error(
+        "Looks like you're trying to enter data for a future date 😅",
+        errorToastOptions
+      );
       return;
     } else if (
       deleteBodyWeight === false &&
@@ -86,18 +114,25 @@ export default function DeleteDataModal(props: any) {
     } else {
       const user_id = (await supabase.auth.getUser()).data.user?.id;
       if (deleteBodyWeight) {
-        await deleteBodyWeightEntry(user_id, date);
+        setMissingBodyWeightEntry(deleteBodyWeightEntry(user_id, date));
       }
       if (deleteBench) {
-        await deletePREntry(user_id, date, "bench");
+        setMissingBenchEntry(await deletePREntry(user_id, date, "bench"));
       }
       if (deleteSquat) {
-        await deletePREntry(user_id, date, "squat");
+        setMissingSquatEntry(await deletePREntry(user_id, date, "squat"));
       }
       if (deleteDeadlift) {
-        await deletePREntry(user_id, date, "deadlift");
+        setMissingDeadliftEntry(await deletePREntry(user_id, date, "deadlift"));
       }
-      document.location.reload();
+      if (
+        missingBodyWeightEntry === 0 &&
+        missingBenchEntry &&
+        missingSquatEntry &&
+        missingDeadliftEntry
+      ) {
+        document.location.reload();
+      }
     }
   }
